move useChannels to kolibri-common
