+++ kolibri/plugins/learn/assets/src/views/SearchResultsGrid.vue
@@ -76,8 +76,8 @@
   import { ref } from 'kolibri.lib.vueCompositionApi';
   import useKResponsiveWindow from 'kolibri-design-system/lib/composables/useKResponsiveWindow';
   import commonCoreStrings from 'kolibri.coreVue.mixins.commonCoreStrings';
+  import SearchChips from 'kolibri-common/components/SearchChips';
   import CopiesModal from './CopiesModal';
-  import SearchChips from './SearchChips';
   import LibraryAndChannelBrowserMainContent from './LibraryAndChannelBrowserMainContent';
 
   export default {
