+++ kolibri/plugins/learn/assets/src/views/TopicsContentPage.vue
@@ -173,9 +173,9 @@
   import GlobalSnackbar from 'kolibri-common/components/GlobalSnackbar';
   import LearningActivityChip from 'kolibri-common/components/ResourceDisplayAndSearch/LearningActivityChip.vue';
   import SidePanelModal from 'kolibri-common/components/SidePanelModal';
+  import useChannels from 'kolibri-common/composables/useChannels';
   import { PageNames, ClassesPageNames } from '../constants';
   import SkipNavigationLink from '../../../../../../kolibri/core/assets/src/views/SkipNavigationLink';
-  import useChannels from '../composables/useChannels';
   import useContentLink from '../composables/useContentLink';
   import useCoreLearn from '../composables/useCoreLearn';
   import useContentNodeProgress from '../composables/useContentNodeProgress';
