+++ kolibri/plugins/learn/assets/test/views/explore-libraries-page.spec.js
@@ -8,7 +8,7 @@ localVue.use(Vuex);
 localVue.use(VueRouter);
 
 jest.mock('../../src/composables/useCardLayoutSpan');
-jest.mock('../../src/composables/useChannels');
+jest.mock('kolibri-common/composables/useChannels');
 jest.mock('../../src/composables/useDevices');
 jest.mock('../../src/composables/useContentLink');
 jest.mock('../../src/composables/usePinnedDevices');
